requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==1.1.1
h11==0.14.0
idna==3.10
msgpack==1.0.2
prompt-toolkit==3.0.20
protobuf==3.19.4
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==2.3.1
Pygments==2.10.0
pynvim==0.4.3
pyvim==3.0.2
six==1.16.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
wcwidth==0.2.5
RPi.GPIO; platform_system=="Linux"
